add example instead test
