+++ crates/sherpa-rs/Cargo.toml
@@ -110,3 +110,7 @@ path = "../../examples/sense_voice.rs"
 [[example]]
 name = "paraformer"
 path = "../../examples/paraformer.rs"
+
+[[example]]
+name = "transducer"
+path = "../../examples/transducer.rs"
