+++ crates/sherpa-rs/src/transducer.rs
@@ -169,54 +169,3 @@ impl Drop for TransducerRecognizer {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::read_audio_file;
-    use std::time::Instant;
-
-    #[test]
-    fn test_transducer_decode() {
-        let path = "motivation.wav";
-        let (samples, sample_rate) = read_audio_file(path).unwrap();
-
-        // Check if the sample rate is 16000
-        if sample_rate != 16000 {
-            panic!("The sample rate must be 16000.");
-        }
-
-        let config = TransducerConfig {
-            decoder: "vosk-en/decoder-epoch-90-avg-20.onnx".to_string(),
-            encoder: "vosk-en/encoder-epoch-90-avg-20.onnx".to_string(),
-            joiner: "vosk-en/joiner-epoch-90-avg-20.onnx".to_string(),
-            tokens: "vosk-en/tokens.txt".to_string(),
-            num_threads: 1,
-            sample_rate: 16_000,
-            feature_dim: 80,
-
-            // NULLs
-            bpe_vocab: "".to_string(),
-            decoding_method: "".to_string(),
-            hotwords_file: "".to_string(),
-            hotwords_score: 0.0,
-            modeling_unit: "".to_string(),
-            blank_penalty: 0.0,
-            debug: true,
-            provider: None,
-        };
-
-        let mut recognizer = TransducerRecognizer::new(config).unwrap();
-
-        let start_t = Instant::now();
-        let result = recognizer.transcribe(sample_rate, &samples);
-        let lower_case = result.to_lowercase();
-        let trimmed = lower_case.trim();
-
-        println!("Time taken for decode: {:?}", start_t.elapsed());
-
-        let expected_result = "the person you can control people are not who you want them to be kill your idols here there are things we can learn from people but people aren't going to be what you think they are what they should";
-
-        assert_eq!(trimmed, expected_result);
-    }
-}
